pyproject.toml
[project]
name = "md-explorer-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.26.0",
    "python-dotenv>=1.2.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.4",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]  # errors, pyflakes, isort

[tool.mypy]
python_version = "3.12"
strict = true
files = ["src/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
